+++ packages/patternfly-4/react-core/src/components/Button/Button.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { css, getModifier } from '@patternfly/react-styles';
 import PropTypes from 'prop-types';
-import { componentShape } from '../../internal/componentShape';
+import { componentShape } from '../../helpers/componentShape';
 import styles from '@patternfly/patternfly/components/Button/button.css';
 
 export const ButtonVariant = {
