+++ packages/patternfly-4/react-core/src/components/Dropdown/DropdownItem.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import styles from '@patternfly/patternfly/components/Dropdown/dropdown.css';
 import { css } from '@patternfly/react-styles';
 import PropTypes from 'prop-types';
-import { componentShape } from '../../internal/componentShape';
+import { componentShape } from '../../helpers/componentShape';
 import { DropdownContext } from './dropdownConstants';
 
 const propTypes = {
