+++ packages/patternfly-4/react-core/src/components/Dropdown/Item.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import DropdownItem from './DropdownItem';
 import styles from '@patternfly/patternfly/components/Dropdown/dropdown.css';
 import { css } from '@patternfly/react-styles';
-import { componentShape } from '../../internal/componentShape';
+import { componentShape } from '../../helpers/componentShape';
 import { DropdownArrowContext } from './dropdownConstants';
 
 const Item = ({ className, ...props }) => (
