+++ packages/patternfly-4/react-core/src/components/Modal/Modal.js
@@ -2,9 +2,9 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import PropTypes from 'prop-types';
 import ModalContent from './ModalContent';
-import safeHTMLElement from '../../internal/safeHTMLElement';
+import safeHTMLElement from '../../helpers/safeHTMLElement';
 import { canUseDOM } from 'exenv';
-import { KEY_CODES } from '../../internal/constants';
+import { KEY_CODES } from '../../helpers/constants';
 import { css } from '@patternfly/react-styles';
 import styles from '@patternfly/patternfly/components/Backdrop/backdrop.css';
