+++ packages/patternfly-4/react-core/src/components/Popover/Popover.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import Tippy from '@tippy.js/react';
 import FocusTrap from 'focus-trap-react';
-import { KEY_CODES } from '../../internal/constants';
+import { KEY_CODES } from '../../helpers/constants';
 import styles from '@patternfly/patternfly/components/Popover/popover.css';
 import { css, getModifier } from '@patternfly/react-styles';
 import PopoverArrow from './PopoverArrow';
@@ -11,7 +11,7 @@ import PopoverBody from './PopoverBody';
 import PopoverHeader from './PopoverHeader';
 import PopoverFooter from './PopoverFooter';
 import PopoverCloseButton from './PopoverCloseButton';
-import GenerateId from '../../internal/GenerateId/GenerateId';
+import GenerateId from '../../helpers/GenerateId/GenerateId';
 import { c_popover_MaxWidth as popoverMaxWidth } from '@patternfly/react-tokens';
 import { tippyStyles } from '../Tooltip/styles';
