+++ packages/patternfly-4/react-core/src/components/Progress/Progress.js
@@ -3,7 +3,7 @@ import styles from '@patternfly/patternfly/components/Progress/progress.css';
 import { css, getModifier } from '@patternfly/react-styles';
 import PropTypes from 'prop-types';
 import ProgressContainer, { ProgressMeasureLocation, ProgressVariant } from './ProgressContainer';
-import { getUniqueId } from '../../internal/util';
+import { getUniqueId } from '../../helpers/util';
 
 export const ProgressSize = {
   sm: 'sm',
