+++ packages/patternfly-4/react-core/src/components/Switch/Switch.js
@@ -2,7 +2,7 @@ import React from 'react';
 import styles from '@patternfly/patternfly/components/Switch/switch.css';
 import { css } from '@patternfly/react-styles';
 import PropTypes from 'prop-types';
-import { getUniqueId } from '../../internal/util';
+import { getUniqueId } from '../../helpers/util';
 
 const propTypes = {
   /** id for the label. */
