+++ packages/patternfly-4/react-core/src/helpers/GenerateId/GenerateId.d.ts
