+++ packages/patternfly-4/react-core/src/helpers/componentShape.js
