+++ packages/patternfly-4/react-core/src/helpers/safeHTMLElement.js
