+++ packages/patternfly-4/react-core/src/helpers/util.ts
