+++ packages/patternfly-4/react-core/src/index.ts
@@ -1,5 +1,5 @@
 export * from './components';
 export * from './layouts';
-export * from './internal';
+export * from './helpers';
 export { GutterSize } from './styles/gutters';
 export { BaseSizes, DeviceSizes } from './styles/sizes';
